- Compile fix for *nix (I hope).
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5494 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Pet.h
@@ -202,7 +202,7 @@ class Pet : public Creature
         time_t  m_resetTalentsTime;
 
         uint64 GetAuraUpdateMask() { return m_auraUpdateMask; }
-        void SetAuraUpdateMask(uint8 slot) { m_auraUpdateMask |= (1i64 << slot); }
+        void SetAuraUpdateMask(uint8 slot) { m_auraUpdateMask |= (uint64(1) << slot); }
         void ResetAuraUpdateMask() { m_auraUpdateMask = 0; }
 
         bool    m_removed;                                  // prevent overwrite pet state in DB at next Pet::Update if pet already removed(saved)
